requirements.txt
python-telegram-bot==20.3
python-dotenv==1.0.0
sqlite3==2.6.0